refactor: extract PDF table source completion
